Add descriptive SEO titles and metadata to radio pages (PSY-331)
Convert radio pages from client-only to server+client split so
generateMetadata() can export dynamic titles and OpenGraph data.
Each page fetches entity data server-side (station/show/episode)
to build descriptive titles following the existing pattern from
artist and show pages.

- /radio -> "Radio | Psychic Homily"
- /radio/[station] -> "KEXP — Radio | Psychic Homily"
- /radio/[station]/[show] -> "Morning Show — KEXP | Psychic Homily"
- /radio/[station]/[show]/[date] -> "Morning Show — Apr 12, 2026 | Psychic Homily"

Closes PSY-331
